incidents_pipeline/requirements.txt
python-dotenv
stomp.py
xmltodict
pytest
pylint